Added service file
The service file uses sockets to communicate with the frontend program.
This service will allow for sudo commands to be ran and sent back to the
program.

It should be noted that this service should not be able to create new
configuration. I don't believe that is a very safe approach, it can be
hard to verify reputability of the source when the file is created.